Initial version of content crawling with a regex...Needs testing

+++ nowcrawling.py
@@ -48,7 +48,7 @@ def getTagsRe(tags):
 def getTypesRe(types):
     return types.replace(' ', '|')
 
-def getURLs(crawlurl, tags, regex2, types):
+def crawlURLs(crawlurl, tags, regex2, types, getfiles):
     url = crawlurl
     headers = {'User-Agent': GOOGLE_USER_AGENT, }
 
@@ -62,23 +62,37 @@ def getURLs(crawlurl, tags, regex2, types):
     except:
         print(BColors.FAIL+'URL '+crawlurl+' not available'+BColors.ENDC)
         return []
-    if (tags is None and regex2 is None):
-        regex = FILE_REGEX.replace('tagholder', '').replace('typeholder', getTypesRe(types))
-    elif (tags is not None):
-        regex = FILE_REGEX.replace('tagholder', getTagsRe(tags)).replace('typeholder', getTypesRe(types))
-    else:
-        regex = FILE_REGEX.replace('tagholder', regex2+'[^<> ]').replace('typeholder', getTypesRe(types))
 
-    ##print(regex)
+    if (getfiles):
+        if (tags is None and regex2 is None):
+            regex = FILE_REGEX.replace('tagholder', '').replace('typeholder', getTypesRe(types))
+        elif (tags is not None):
+            regex = FILE_REGEX.replace('tagholder', getTagsRe(tags)).replace('typeholder', getTypesRe(types))
+        else:
+            regex = FILE_REGEX.replace('tagholder', regex2+'[^<> ]').replace('typeholder', getTypesRe(types))
+    else:
+        ##FIXME CONTENT BLA BLA BLA
+        regex = regex2
+
     p = re.compile(regex, re.IGNORECASE)
 
     tuples = p.findall(data)
+
     if (len(tuples) < 1):
         return []
-    tuples = [j[i] for j in tuples for i in range(len(j)) if '.' in j[i]]
 
-    prettyurls = list(x.replace('href=', '').replace('HREF=', '').replace('"', '').replace('\\', '') for x in tuples)
-    prettyurls = list(crawlurl+x if "://" not in x else x for x in prettyurls)
+    if (getfiles):
+        tuples = [j[i] for j in tuples for i in range(len(j)) if '.' in j[i]]
+        prettyurls = list(x.replace('href=', '').replace('HREF=', '').replace('"', '').replace('\\', '') for x in tuples)
+        prettyurls = list(crawlurl+x if "://" not in x else x for x in prettyurls)
+    else:
+        ##FIXME CONTENT BLA BLA BLA
+        ## RETURN EVERYTHING IF TUPLES
+        if (isinstance(tuples[0],tuple)):
+            prettyurls = list(j[i] for j in tuples for i in range(len(j)))
+        else:
+            prettyurls = list(x for x in tuples)
+
     return prettyurls
 
 def parse_input():
@@ -86,9 +100,9 @@ def parse_input():
     parser = OptionParser()
     parser.add_option('-f', '--files', help='Crawl for files', action="store_true", dest="getfiles")
     parser.add_option('-c', '--content', help='Crawl for content (words, strings, pages, regexes)', action="store_false", dest="getfiles")
+    parser.add_option('-k', '--keywords', help='(Required) A quoted list of words separated by spaces which will be the search terms of the crawler', dest='keywords', type='string')
 
     filesgroup = OptionGroup(parser, "Files (-f) Crawler Arguments")
-    filesgroup.add_option('-k', '--keywords', help='(Required) A quoted list of words separated by spaces which will be the search terms of the crawler', dest='keywords', type='string')
     filesgroup.add_option('-a', '--ask', help='Ask before downloading', action="store_true", dest="ask", default=False)
     filesgroup.add_option('-l', '--limit', help='File size limit in bytes separated by a hyphen (example: 500-1200 for files between 500 and 1200 bytes, -500 for files smaller than 500 bytes, 500- for files larger than 500 bytes) (Default: None)', dest="limit", type='string', default=None)
     filesgroup.add_option('-n', '--number', help='Number of files to download until crawler stops (Default: Max)', dest="maxfiles", type='int', default=None)
@@ -102,29 +116,25 @@ def parse_input():
     parser.add_option_group(filesgroup)
     parser.add_option_group(filenamegroup)
 
-    ##contentgroup = OptionGroup(parser, "Content Crawler Options","When crawling for content, the following parameters are available:")
-    ##contentgroup.add_option("-c", action="store_true", help="Crawl for Content.")
-    ##parser.add_option_group(contentgroup)
+    contentgroup = OptionGroup(parser, "Content (-c) Crawler Arguments")
+    contentgroup.add_option('-m', '--match', help='(Required) A regex that will match the content you are crawling for', dest='regex', type='string')
 
-    ##filesgroup = OptionGroup(parser, "Debug Options")
-    ##filesgroup.add_option('-t', '--tags', help='A quoted list of words separated by spaces that must be present in the file name or content that you\'re crawling for', dest='tags', type='string', default='')
-    ##parser.add_option_group(filesgroup)
+    parser.add_option_group(contentgroup)
 
-
-    ##parser.add_option('-c', '--content', help='Crawl for content (words, strings, pages) instead of files', action="store_false", dest="getfiles", default=True)
-    ##parser.add_option('-t', '--tags', help='A quoted list of words separated by spaces that must be present in the file name or content that you\'re crawling for', dest='tags', type='string', default='')
-    ##parser.add_option('-r', '--regex', help='Instead of tags you can just specify a regex for the content you\'re looking for', dest='regex', type='string', default='')
-    ##parser.add_option('-h', '--hint', help='A quoted list of words separated by spaces which will be the search terms of the crawler (Default will be tags)', dest='hint', type='string', default='')    ##parser.add_option('-c', '--connection-timeout', help='Set the connection timeout waiting time (default: 60)', dest='connection_timeout', type='float', default=60)
     (options, args) = parser.parse_args()
 
     if (options.getfiles is None):
         parser.error('You must specify the crawler type: -f for files or -c for content')
     if (options.getfiles is True and options.keywords is None):
         parser.error('You must specify keywords when crawling for files.')
+    if (options.getfiles is False and options.keywords is None):
+        parser.error('You must specify keywords when crawling for content.')
     if (options.getfiles is True and (options.tags is not None and options.regex is not None)):
         parser.error('You can\'t pick both file name search options: -t or -r')
     if (options.getfiles is True and options.limit is not None and '-' not in options.limit):
         parser.error('Limits must be separated by a hyphen.')
+    if (options.getfiles is False and options.regex is None):
+        parser.error('You must specify a matching regex (-m) when crawling for content.')
     ## FIXME FALTA TANTO CHECK AI JASUS, TIPO VER SE O GAJO NAO METE -A SEM METER -F ENTRE OUTROS AI JASUS
 
     return options.getfiles, options.keywords, options.extensions, options.smart, options.tags, options.regex, options.ask, options.limit, options.maxfiles
@@ -151,36 +161,41 @@ def crawl(getfiles, keywords, extensions, smart, tags, regex, ask, limit, maxfil
             googleurls = crawlGoogle(GOOGLE_NUM_RESULTS, start, keywords, smart)
             print(len(googleurls))
             for searchurl in googleurls:
-                downloadurls = getURLs(searchurl, tags, regex, extensions)
+                downloadurls = crawlURLs(searchurl, tags, regex, extensions, getfiles)
                 urllib.request.urlcleanup()
                 if len(downloadurls) < 1:
                     print('No results in '+searchurl)
                 else:
-                    for file in downloadurls:
-                        if (maxfiles is not None and downloaded >= maxfiles):
-                            print(BColors.OKGREEN+'All files have been downloaded. Exiting...'+BColors.ENDC)
-                            exit()
-                        try:
-                            filename = file.split('/')[-1]
-                            meta = urllib.request.urlopen(file).info()
-                            filesize = meta.get_all("Content-Length")[0]
-                            if (limit is not None and (maxsize is not None and (int(filesize) > int(maxsize)) or (minsize is not None and (int(filesize) < int(minsize))))):
-                                print (BColors.WARNING+'Skipping file '+filename+' because size '+filesize+' is off limits.'+BColors.ENDC)
+                    if (getfiles):
+                        for file in downloadurls:
+                            if (maxfiles is not None and downloaded >= maxfiles):
+                                print(BColors.OKGREEN+'All files have been downloaded. Exiting...'+BColors.ENDC)
+                                exit()
+                            try:
+                                filename = file.split('/')[-1]
+                                meta = urllib.request.urlopen(file).info()
+                                filesize = meta.get_all("Content-Length")[0]
+                                if (limit is not None and (maxsize is not None and (int(filesize) > int(maxsize)) or (minsize is not None and (int(filesize) < int(minsize))))):
+                                    print (BColors.WARNING+'Skipping file '+filename+' because size '+filesize+' is off limits.'+BColors.ENDC)
+                                    continue
+                                if ask:
+                                    print (BColors.OKBLUE+'Download file '+filename+' of size '+filesize+' from '+file+'? [y/n]:'+BColors.ENDC)
+                                    choice = input().lower()
+                                    if choice not in YES:
+                                       continue
+                                print (BColors.OKGREEN+'Downloading file '+filename+' of size '+filesize+BColors.ENDC)
+                                urllib.request.urlretrieve(file, filename)
+                                downloaded += 1
+                            except KeyboardInterrupt:
+                                print(BColors.FAIL+'Interrupted. Exiting...'+BColors.ENDC)
+                                exit()
+                            except:
+                                print(BColors.FAIL+'File '+file+ ' from '+searchurl+' not available'+BColors.ENDC)
                                 continue
-                            if ask:
-                                print (BColors.OKBLUE+'Download file '+filename+' of size '+filesize+' from '+file+'? [y/n]:'+BColors.ENDC)
-                                choice = input().lower()
-                                if choice not in YES:
-                                   continue
-                            print (BColors.OKGREEN+'Downloading file '+filename+' of size '+filesize+BColors.ENDC)
-                            urllib.request.urlretrieve(file, filename)
-                            downloaded += 1
-                        except KeyboardInterrupt:
-                            print(BColors.FAIL+'Interrupted. Exiting...'+BColors.ENDC)
-                            exit()
-                        except:
-                            print(BColors.FAIL+'File '+file+ ' from '+searchurl+' not available'+BColors.ENDC)
-                            continue
+                    else:
+                        for match in downloadurls:
+                            ##FIXME CONTENT BLA BLA
+                            print(BColors.OKGREEN+match+BColors.ENDC)
             if len(googleurls) < GOOGLE_NUM_RESULTS:
                 break
             else:
